requirements.txt

psycopg2
fastapi==0.104.1              # Web framework
gunicorn==22.0.0
uvicorn[standard]==0.24.0     # ASGI server


sqlalchemy==2.0.23            # ORM
alembic==1.13.0               # Database migrations
python-dotenv==1.0.0          # Environment variables
sse-starlette==2.1.0


cloudinary
pydantic            # Schema validation
pydantic-settings==2.10.1      # Settings management
python-multipart==0.0.6       # File upload handling
pydantic[email]

# Compatible triple for LangChain 1.0+
# (classic chains and memory are now in langchain-classic)
langchain==1.0.0              # LLM orchestration framework
langchain-core==1.0.2
langchain-community==0.4.1
langchain-google-genai==3.0.0 # Gemini integration
langchain-openai
langchain_text_splitters

qdrant-client==1.15.1
sentence-transformers==5.1.1

python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4            # Password hashing


python-dateutil==2.8.2        # Date/time utilities


PyPDF2==3.0.1                 # Extract text from PDF
pdfplumber==0.10.3            # Better PDF parsing (Vietnamese support)
python-docx==0.8.11           # Extract text from DOCX/DOC
python-pptx==0.6.21           # Extract text from PPTX (optional)
openpyxl==3.1.2               # Extract text from XLSX (optional)
beautifulsoup4==4.12.2        # HTML parsing
lxml==4.9.3                   # XML parsing (needed by BeautifulSoup)


unidecode==1.3.7              # Handle Vietnamese characters/accents



python-json-logger==2.0.7     # JSON logging for production
